Temporary solution to work around stupid error of config.ini file (could not be run more than 1020 times)

+++ src/dev/hpetersson/12BarTensegrity/T12Controller.h
@@ -135,7 +135,7 @@ class T12Controller : public tgObserver<T12Model>
 	void saveData2File(void);
 
 	void clearParams(void);
-	int simulationNumber;
+	double simulationNumber;
 };
 
 #endif // T12CONTROLLER
